+++ common/install-timers.sh
@@ -98,7 +98,6 @@ UNIT
 ws_write_unit "$UNIT_DIR/clamav-scan.service" <<UNIT
 [Unit]
 Description=ClamAV dagelijkse scan
-After=clamav-freshclam.service
 
 [Service]
 Type=oneshot